优化job project任务

+++ streamx-console/streamx-console-service/src/main/java/com/streamxhub/streamx/console/core/service/impl/ProjectServiceImpl.java
@@ -128,7 +128,10 @@ public class ProjectServiceImpl extends ServiceImpl<ProjectMapper, Project>
             return false;
         }
         try {
-            project.delete();
+            if (project.getRepository() != 3)
+            {
+                project.delete();
+            }
             removeById(id);
             return true;
         } catch (IOException e) {
